+++ lib/features/general/custom_chip.widget.dart
@@ -3,12 +3,14 @@ import 'package:flutter/material.dart';
 class CustomChip extends StatelessWidget {
   final Text label;
   final Color? color;
+  final Widget? icon;
   final EdgeInsets padding;
 
   const CustomChip({
     Key? key,
     required this.label,
     this.color,
+    this.icon,
     this.padding = const EdgeInsets.symmetric(
       vertical: 1,
       horizontal: 6,
@@ -26,7 +28,13 @@ class CustomChip extends StatelessWidget {
         child: Container(
           color: color ?? Theme.of(context).chipTheme.backgroundColor,
           padding: padding,
-          child: label,
+          child: Row(
+            mainAxisSize: MainAxisSize.min,
+            children: [
+              if (icon != null) ...[icon!, const SizedBox(width: 5)],
+              label,
+            ],
+          ),
         ),
       ),
     );